Adds Authorization via SNAC role for Vocabulary Editors
Using the SNAC tables that keep track of roles we are now validating
that the current user belongs to the role that has the correct
capability to edit Vocabulary. This was done implementing a new
permission Gate ('edit-vocabulary') that can be used in blade templates
as in other places where authorization is required.